feat(Admin): admin 정보 수정 추가

+++ src/main/java/com/esc/bluespring/domain/member/MemberController.java
@@ -6,16 +6,17 @@ import static com.esc.bluespring.domain.member.entity.Member.STUDENT;
 import com.esc.bluespring.common.security.CustomJwtEncoder;
 import com.esc.bluespring.common.utils.file.S3Service;
 import com.esc.bluespring.domain.file.entity.Image;
-import com.esc.bluespring.domain.member.classes.AdminMapper;
 import com.esc.bluespring.domain.member.classes.MemberDto.AdminJoin;
 import com.esc.bluespring.domain.member.classes.MemberDto.Join;
 import com.esc.bluespring.domain.member.classes.MemberDto.JwtToken;
 import com.esc.bluespring.domain.member.classes.MemberDto.Login;
 import com.esc.bluespring.domain.member.classes.MemberDto.Patch;
 import com.esc.bluespring.domain.member.classes.MemberDto.SendFriendShipRequest;
-import com.esc.bluespring.domain.member.classes.StudentMapper;
+import com.esc.bluespring.domain.member.classes.MemberMapper;
+import com.esc.bluespring.domain.member.entity.Admin;
 import com.esc.bluespring.domain.member.entity.Member;
 import com.esc.bluespring.domain.member.entity.Student;
+import com.esc.bluespring.domain.member.exception.MemberException.MemberNotFoundException;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.servlet.http.Cookie;
@@ -39,8 +40,7 @@ public class MemberController {
 
     private final MemberServiceFacade memberServiceFacade;
     private final S3Service s3Service;
-    private final StudentMapper mapper = StudentMapper.INSTANCE;
-    private final AdminMapper adminMapper = AdminMapper.INSTANCE;
+    private final MemberMapper mapper = MemberMapper.INSTANCE;
     private final CustomJwtEncoder customJwtEncoder;
 
     @PostMapping("join")
@@ -54,24 +54,31 @@ public class MemberController {
     @PostMapping("admin-join")
     @ResponseStatus(HttpStatus.CREATED)
     public void join(@RequestBody AdminJoin dto) {
-        memberServiceFacade.join(adminMapper.toEntity(dto));
+        memberServiceFacade.join(mapper.toEntity(dto));
     }
 
     @DeleteMapping("resign")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @RolesAllowed({ADMIN, STUDENT})
-    public void resign(Student student) {
-        s3Service.remove(student.getProfileImage().getUrl());
-        memberServiceFacade.resign(student);
+    public void resign(Member member) {
+        if (member instanceof Student student) {
+            s3Service.remove(student.getProfileImage().getUrl());
+        }
+        memberServiceFacade.resign(member);
     }
 
     @PostMapping("{id}/friendship-requests")
     @ResponseStatus(HttpStatus.CREATED)
+    @RolesAllowed({STUDENT})
     @Operation(description = "친구신청을 합니다.")
-    public void requestFriendship(Member user, @PathVariable Long id,
+    public void requestFriendship(Student user, @PathVariable Long id,
         @RequestBody SendFriendShipRequest dto) {
-        Member target = memberServiceFacade.find(id);
-        memberServiceFacade.requestFriendship(user, target, dto.message());
+        try {
+            Student target = (Student) memberServiceFacade.find(id);
+            memberServiceFacade.requestFriendship(user, target, dto.message());
+        } catch (ClassCastException e) {
+            throw new MemberNotFoundException();
+        }
     }
 
     @PostMapping("login")
@@ -86,9 +93,17 @@ public class MemberController {
 
     @PatchMapping("/me")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void patch(@RequestBody @Valid Patch dto, Student student) {
-        Image profile = s3Service.upload(dto.profileImage());
-        Image certificateImage = s3Service.upload(dto.studentCertificationImage());
-        memberServiceFacade.patch(mapper.toEntity(dto, profile, certificateImage), student);
+    public void patch(@RequestBody @Valid Patch dto, Member target) {
+        if (target instanceof Admin) {
+            Admin source = mapper.toEntity(dto);
+            memberServiceFacade.patch(source, target);
+            return;
+        }
+        if (target instanceof Student) {
+            Image profile = s3Service.upload(dto.profileImage());
+            Image certificateImage = s3Service.upload(dto.studentCertificationImage());
+            Student source = mapper.toEntity(dto, profile, certificateImage);
+            memberServiceFacade.patch(source, target);
+        }
     }
 }
